requirements.txt
openai>=1.0.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
requests>=2.28.0
urllib3>=1.26.0
charset-normalizer>=2.0.0
idna>=3.3
certifi>=2022.0.0
python-dotenv>=0.19.0
openpyxl>=3.0.0
XlsxWriter>=3.0.0
PyInstaller>=5.0.0